Simplified how BitSlices compute fast path.
In particular, BitSlice is now four words instead of three. It uses
the extra word to hold the number of full, aligned blocks at the
beginning of the slice. For unaligned slices, this is 0. In either
case, an initial bounds check against `full_blocks` lets us access
a block with only one branch, because if the check succeeds then
we know not only are we in bounds, but the offset points to a full
block.